Add 1 star and 1 planet
Add some basic information, but just a small part, of HR5756 and Ikrinas.

HR5756 is the main star of the system in this mod.
Ikrinas is a planet with life, like our earth. Kerbals (maybe will address them as something else) will live on it.

I havn't made textures for them.

Notes: The original Orbital Period of Ikrinas is about 2.50~2.51 years (the Earth year). It will be changed more accurate according to some simulations.
Its original Rotational Period is 91542 seconds. It may be changed easier to remember in the future.

These files are recovered : (
I'm not simular to Github Desktop, and delete them by a mistake!
I can't find the localization.cfg forever!!!!